Fix some inference problems
Inference seems to sometimes have a hard time with types-as-arguments unless they are declared as ::Type{T}. Since we were using Vararg for the border, this meant that many operations were not type stable. This switches to a border-instance rather than border-type specification.

This also makes factorkernel type-stable, at the cost of an extra copy of image data.