Fix inference of imgradients
Earlier versions of Julia supported two types of lispy tuple programming:
one in which arguments shrink (e.g., via Base.tail) and the output gets
appended, and the other in which the result is appended to one of the
arguments (and therefore grows). For 0.7, inference has prioritized
either (1) proving that inference will converge or (2) exiting quickly.
Proving convergence is apparently much harder if some of the arguments
grow in size, so inference doesn't try very hard to get the right ansswer
anymore. Consequently inference was broken in our previous implementation.

This slight disadvantage is more than compensated by the fact that we
can now do inferrable arithmetic on type parameters, as exemplified here
by `ReshapedOneD{N,N-M}(arg)`.