Restrict depwarn to non-concrete color types (#137)
The only behavior that's changing is `convert(Array{RGB}, A)`, which
will no longer choose a numeric type `T` for `RGB{T}`---instead it
will literally be an `Array{RGB}`. But `convert(Array{RGB{Float32}}, A)`
will continue to work as before (it will just rely on the Base machinery),
so don't issue a warning.