add strel_size
Typically, a fast window filter implementation separates the loop into
two parts: 1) inner loop where bounds check are unnecessary, and 2)
boundary loop where one has to carefully deal with boundary conditions.

This helper function can be used to infer the minimal radius of the
strucuring element so that separating the loop becomes an easy work with
the help of TiledIteration.